requirements.txt
av==14.0.0
dataclasses_json==0.6.7
decord==0.6.0
einops==0.8.0
Flask==3.1.0
flask_cors==5.0.0
gradio==5.8.0
h5py==3.12.1
huggingface_hub==0.24.5
hydra-core==1.3.2
imageio==2.35.1
imagesize==1.4.1
iopath==0.1.10
kapture==1.1.10
kapture_localization==1.1.10
matplotlib==3.9.3
numpy==2.1.3
omegaconf==2.3.0
open3d==0.18.0
opencv_python==4.10.0.84
packaging==24.2
pandas==2.2.3
Pillow==11.0.0
pillow_heif==0.21.0
plyfile==1.1
poselib==2.0.4
pycocotools==2.0.8
pycolmap==3.11.0
pyrender==0.1.45
quaternion==3.5.2.post4
roma==1.5.0
scikit_learn==1.5.2
scipy==1.14.1
setuptools==72.1.0
skimage==0.0
splines==0.3.2
strawberry==3.0
submitit==1.5.2
tensordict==0.6.2
tqdm==4.66.5
trimesh==4.5.3
waymo_open_dataset==1.0.1
